feat: SSR toolbar
